Update trie.c
seems like a big part of the problem is that my function is set up to return a pointer. SO maybe it's a design error. 
It's kind of a catch 22. I want to return a pointer that I'm allocating memory for, but I need to also free that memory. I bet its possible to finagle my way through that situation, but my intuition tells me that if I'm finagling for something that seems this basic, I'm going about it wrong.
I will do a little more research to confirm that a redesign is my best path, but as of right now it looks like it.