Update trie.c
Ok, so I watched a bunch of info on valgrind and read multiple articles and I'm still kinda lost. But one idea that I got was that for every malloc there should be a free. Prety basic right? But i guess I overlooked the fact that theres like a thousand mallocs as I'm storing all these words and only one free in unload. Even though I guess it makes sense that if I free the head node, all the rest should be freed, that's not what happens.

Hmmm, I guess now that I mention it, it doesnt make sense. Cause the other words would still be malloced on the head. HMMMM. Well to be honest I tried fixing this problem in the hashtable by  introducing a for loop or something that iterates through all the buckets and frees the ones that have something in them. I remember valgrind still complaining though. But to be fair, that code is way more mangy and twisted up so it could've been some other thing messing me up. 

Let's try individually freeing everything.