Update trie.c
so I did a bunch more research on the correct way to malloc in general and then more specifically structs, and i think I've got a bunch of stuff right. I have been a bit redundant in specific areas which I will label, which is why when I was messing around, I could remove some init_nodes and have the program work fine. Now back to writing the code to free everything